feat(PriceAPI, TxErrorCode, EssValue) : 요금 관련 api, 트랜잭션 종료에 대한 오류 코드 저장 및 조회, ess값 차감 및 조회

+++ src/main/java/com/example/quaterback/api/domain/station/entity/ChargingStationEntity.java
@@ -33,7 +33,7 @@ public class ChargingStationEntity {
     private Double longitude;
     private String address;
     private LocalDateTime updateStatusTimeStamp;
-    private Integer essValue;
+    private Double essValue;
 
     @Enumerated(EnumType.STRING)
     private StationStatus stationStatus;
@@ -61,7 +61,7 @@ public class ChargingStationEntity {
         updateStatusTimeStamp = LocalDateTime.now();
     }
 
-    public void updateStationEssValue(Integer value){
+    public void updateStationEssValue(Double value){
         essValue = value;
     }
