+++ src/test/java/com/example/quaterback/websocket/transaction/event/fixture/TransactionEventFixture.java
@@ -183,7 +183,7 @@ public class TransactionEventFixture {
             String transactionId,
             Integer id,
             String idToken,
-            Integer value
+            Double value
     ) {
         return new TransactionEventDomain(
                 messageTypeId,
